fix regression: expand ~/ before storing search directory in query object
The regression was introduced when the SearchQuery object was added
and moved up some lines; as result, a path starting with tilde character
would not be expanded to the user's home directory, and the search
results would be broken.

This fix now has the minor quirk that when searching again, the
directory text field in search dialog will contain the expanded path
instead of the original one entered by the user.