normalize directory paths (normpath + add trailing slash)
Normalizing the directory path before displaying search dialog and
before starting search prevents the list of recent dirs from being
filled with nearly-duplicate entries.
I've decided to add a trailing slash for normalization, because
the paths are expected to be directories anyway, so it shouldn't
make a difference in case of symlinks.