make sure displayed file names are correctly UTF8-encoded
It is possible that files names are not valid UTF8 (for
example because they are in ISO-8859-1), but are
still interpreted as UTF8. For these cases, the displayed
names must be converted to correct encoding before
being displayed (otherwise Pango will complain).

The gobject.filename_display_name() method is here
preferred over Pythons unicode() method so we get
the same displayed name as shown in Gedit.

(To follow the reasoning, create a file with German word
"rückgang" as name, encoded in ISO-8859-1, and
then interpret the file name as UTF8).