bugfix in ProcessInfo: handle IOError when opening process status file
If a status file cannot be opened, the process is ignored
(it is not added to the list of processes at all).

Such an error can occur if a process ends while we go
through the list of PIDs. To reproduce, run something
like "while [ /bin/true ]; do /bin/true; done" in another shell
and then try to Cancel a search. Without this fix, an error
might occur that prevents the cancelling.