refactor SearchProcess and GrepParser into several new classes
We now have SearchProcess class for creating the search
command line (find/xargs/grep) and parsing the grep results,
then RunCommand class for running the command in background
and catching its output, and LineSplitter class for splitting
the output into lines which are fed back to SearchProcess.

This should make it easier to start find and grep as
separate commands later.