use directory specified by ProjectMarker plugin, if available
If the ProjectMarker plugin has set a valid project root
directory for the current file, this directory is used as
base directory for searching.

Note: if a search has already been made in a
different directory, that previous directory will still be
used as base dir. It would be better if the "last search directory"
would be stored for every project, but that's probably a task
for a full project management plugin.

Note 2: so far, the current ProjectMarker plugin creates
paths with trailing newlines, which will not work as base dirs...