add "Whole word" search option
Although grep already offers a -w flag, it's behavior doesn't match
what I want (esp. when the search text itself begins or ends with a
non-word character). So this implementation now performs a separate Regex
search on the grep results.
The idea how to build this Regex is taken from the "ack" search tool
and IMO gives good results.