add test file for GNU Grep bug regarding Turkish Unicode characters
The test file contains an upper-case I with a dot (U+0130). When searching
for "a" with case-insensitive search, some GNU Grep versions will omit the
newline between the two lines in this file, which will cause incorrect
output in the plugin's result panel.

Grep 2.10 under Ubuntu 12.04 has this bug; Grep 2.14 from upstream doesn't
have it. I don't see any good way to work around it in this plugin, but I
guess the impact should be quite minor anyway.