reduce memory leak in GtkTreeStore
Looks like there's a memory leak when using Gtk.TreeStore.append() with a
list of initial values. I've seen this at least on Ubuntu 12.04. Might be
bugs.gnome.org/693402 .

Even with this workaround, resident memory usage doesn't go down when
closing a result panel; but at least that memory is reused when starting
a new search. Also there appear to be other memory leaks in pygobject
under Ubuntu 12.04, which are fixed eg. in Fedora 19.

Test:
- start Gedit, and watch its RES memory usage in "top"
- start a huge search (eg. search for "a" in /usr/share)
- stop search when ca. 300.000 results have been added
- note RES memory usage of gedit process
- close result pane
- start same search again, and cancel it at ca. 300.000 results again
- compare current RES memory usage with previous value
  - good: memory usage has increased by max. 5MB
  - bad: memory usage has increased by more than 20MB