python3 port: adjust file-search.plugin file to request new loader
To provide the same plugin for Python 2 and 3 (ie. Gedit before 3.8 and
Gedit >= 3.8), a second .plugin file has been added now, which provides
the plugin for Gedit < 3.8 (by specifying "Loader=python"). This means
the Gedit plugin selection dialog will now show two File Search plugins,
and the user has to select the correct one for his version.

This is awkward, but I didn't find a better solution so far.